refactor(manager/pip-compile): Move topological sort to utils (#27398)
