+++ lib/modules/manager/pip-compile/utils.ts
@@ -1,4 +1,37 @@
+import { Graph } from 'graph-data-structure';
+import type { PackageFile } from '../types';
 import type { DependencyBetweenFiles, PipCompileArgs } from './types';
+
+export function sortPackageFiles(
+  depsBetweenFiles: DependencyBetweenFiles[],
+  packageFiles: Map<string, PackageFile>,
+): PackageFile[] {
+  const result: PackageFile[] = [];
+  const graph: ReturnType<typeof Graph> = Graph();
+  depsBetweenFiles.forEach(({ sourceFile, outputFile }) => {
+    graph.addEdge(sourceFile, outputFile);
+  });
+  const sorted = graph.topologicalSort();
+  for (const file of sorted) {
+    if (packageFiles.has(file)) {
+      const packageFile = packageFiles.get(file)!;
+      const sortedLockFiles = [];
+      // TODO(not7cd): this needs better test case
+      for (const lockFile of packageFile.lockFiles!) {
+        if (sorted.includes(lockFile)) {
+          sortedLockFiles.push(lockFile);
+        }
+      }
+      packageFile.lockFiles = sortedLockFiles;
+      result.push(packageFile);
+    }
+  }
+  // istanbul ignore if: should never happen
+  if (result.length !== packageFiles.size) {
+    throw new Error('Topological sort failed to include all package files');
+  }
+  return result;
+}
 
 export function generateMermaidGraph(
   depsBetweenFiles: DependencyBetweenFiles[],
